move device plugin manager into resources module
