+++ crates/kubelet/src/lib.rs
@@ -113,13 +113,13 @@ pub(crate) mod mio_uds_windows;
 pub mod backoff;
 pub mod config;
 pub mod container;
-pub mod device_plugin_manager;
 pub mod handle;
 pub mod log;
 pub mod node;
 pub mod plugin_watcher;
 pub mod pod;
 pub mod provider;
+pub mod resources;
 pub mod secret;
 pub mod state;
 pub mod store;
